refactoring keystore and fskeystore tests adding validateName tests
